+++ CHANGES.rst
@@ -12,6 +12,8 @@ Fixes:
 
 - Fix errors when cutting and copying objects in folder contents
   [vangheem]
+- Fixed error message unicode error in rename action.
+  [Gagaro]
 
 
 3.0.16 (2016-01-08)